+++ agent/containermetadata/parse_metadata_test.go
@@ -79,7 +79,11 @@ func TestParseHasConfig(t *testing.T) {
 	mockContainerInstanceARN := containerInstanceARN
 
 	mockConfig := &docker.Config{Image: "image"}
-	mockContainer := &docker.Container{Config: mockConfig}
+
+	mockNetworks := make(map[string]docker.ContainerNetwork)
+	mockNetworkSettings := &docker.NetworkSettings{Networks: mockNetworks}
+
+	mockContainer := &docker.Container{Config: mockConfig, NetworkSettings: mockNetworkSettings}
 
 	expectedStatus := string(MetadataReady)
 
@@ -89,6 +93,7 @@ func TestParseHasConfig(t *testing.T) {
 	}
 
 	metadata := newManager.parseMetadata(mockContainer, mockTaskARN, mockContainerName)
+
 	assert.Equal(t, metadata.cluster, mockCluster, "Expected cluster "+mockCluster)
 	assert.Equal(t, metadata.taskMetadata.containerName, mockContainerName, "Expcted container name "+mockContainerName)
 	assert.Equal(t, metadata.taskMetadata.taskARN, mockTaskARN, "Expcted task ARN "+mockTaskARN)
@@ -97,7 +102,7 @@ func TestParseHasConfig(t *testing.T) {
 	assert.Equal(t, metadata.dockerContainerMetadata.imageName, "image", "Expected nonempty imageID")
 }
 
-func TestParseHasHostConfigPortBindsError(t *testing.T) {
+func TestParseHasNetworkSettingsPortBindings(t *testing.T) {
 	mockTaskARN := validTaskARN
 	mockContainerName := containerName
 	mockCluster := cluster
@@ -105,10 +110,15 @@ func TestParseHasHostConfigPortBindsError(t *testing.T) {
 
 	mockPorts := make(map[docker.Port][]docker.PortBinding)
 	mockPortBinding := make([]docker.PortBinding, 0)
-	mockPortBinding = append(mockPortBinding, docker.PortBinding{})
+	mockPortBinding = append(mockPortBinding, docker.PortBinding{HostIP: "0.0.0.0", HostPort: "8080"})
 	mockPorts["80/tcp"] = mockPortBinding
-	mockHostConfig := &docker.HostConfig{PortBindings: mockPorts}
-	mockContainer := &docker.Container{HostConfig: mockHostConfig}
+
+	mockHostConfig := &docker.HostConfig{NetworkMode: "bridge"}
+	mockNetworks := make(map[string]docker.ContainerNetwork)
+	mockNetworks["bridge"] = docker.ContainerNetwork{}
+	mockNetworks["network0"] = docker.ContainerNetwork{}
+	mockNetworkSettings := &docker.NetworkSettings{Networks: mockNetworks, Ports: mockPorts}
+	mockContainer := &docker.Container{HostConfig: mockHostConfig, NetworkSettings: mockNetworkSettings}
 
 	expectedStatus := string(MetadataReady)
 
@@ -123,36 +133,12 @@ func TestParseHasHostConfigPortBindsError(t *testing.T) {
 	assert.Equal(t, metadata.taskMetadata.taskARN, mockTaskARN, "Expcted task ARN "+mockTaskARN)
 	assert.Equal(t, metadata.containerInstanceARN, mockContainerInstanceARN, "Expcted container instance ARN "+mockContainerInstanceARN)
 	assert.Equal(t, string(metadata.metadataStatus), expectedStatus, "Expected status "+expectedStatus)
-	assert.Equal(t, len(metadata.dockerContainerMetadata.ports), 0, "Expected empty list of ports")
-}
+	assert.Equal(t, len(metadata.dockerContainerMetadata.networkInfo.networks), 2, "Expected two networks")
 
-func TestParseHasHostConfigPortBindsNoError(t *testing.T) {
-	mockTaskARN := validTaskARN
-	mockContainerName := containerName
-	mockCluster := cluster
-	mockContainerInstanceARN := containerInstanceARN
-
-	mockPorts := make(map[docker.Port][]docker.PortBinding)
-	mockPortBinding := make([]docker.PortBinding, 0)
-	mockPortBinding = append(mockPortBinding, docker.PortBinding{HostIP: "0.0.0.0", HostPort: "80"})
-	mockPorts["80/tcp"] = mockPortBinding
-	mockHostConfig := &docker.HostConfig{PortBindings: mockPorts}
-	mockContainer := &docker.Container{HostConfig: mockHostConfig}
-
-	expectedStatus := string(MetadataReady)
-
-	newManager := &metadataManager{
-		cluster:              mockCluster,
-		containerInstanceARN: mockContainerInstanceARN,
-	}
-
-	metadata := newManager.parseMetadata(mockContainer, mockTaskARN, mockContainerName)
-	assert.Equal(t, metadata.cluster, mockCluster, "Expected cluster "+mockCluster)
-	assert.Equal(t, metadata.taskMetadata.containerName, mockContainerName, "Expcted container name "+mockContainerName)
-	assert.Equal(t, metadata.taskMetadata.taskARN, mockTaskARN, "Expcted task ARN "+mockTaskARN)
-	assert.Equal(t, metadata.containerInstanceARN, mockContainerInstanceARN, "Expcted container instance ARN "+mockContainerInstanceARN)
-	assert.Equal(t, string(metadata.metadataStatus), expectedStatus, "Expected status "+expectedStatus)
-	assert.Equal(t, len(metadata.dockerContainerMetadata.ports), 1, "Expected empty list of ports")
+	assert.Equal(t, len(metadata.dockerContainerMetadata.ports), 1, "Expected nonempty list of ports")
+	assert.Equal(t, uint16(80), metadata.dockerContainerMetadata.ports[0].ContainerPort, "Expected nonempty ContainerPort field")
+	assert.Equal(t, uint16(8080), metadata.dockerContainerMetadata.ports[0].HostPort, "Expected nonempty HostPort field")
+	assert.Equal(t, "0.0.0.0", metadata.dockerContainerMetadata.ports[0].BindIP, "Expected nonempty HostIP field")
 }
 
 func TestParseHasNetworkSettingsNetworksEmpty(t *testing.T) {