+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Changelog
+#Changelog
+
+## Unreleased
+
+## 1.15.1
+* Bug - Fixed a bug where a task can be blocked in creating state. [#1048](https://github.com/aws/amazon-ecs-agent/pull/1048)
+* Bug - Fixed dynamic HostPort in container metadata. [#1052](https://github.com/aws/amazon-ecs-agent/pull/1052)
 
 ## 1.15.0
 * Feature - Support for provisioning tasks with ENIs.